Module csil_schedule_api::api

Functions

cover_shift

Covers the shift with the given id

dedup_cleaning 🔒

Gets rid of duplicate CleaningLocations

delete_shift

Deletes and returns the shift with the given id

get_intervals_that_need_cover 🔒

Find the time ranges that would need a cover if ignore_shift was removed

get_shift

Returns the shift with the given id

get_shifts

Gets shifts based on a ShiftQuery

get_shifts_json

Gets shifts based on a ShiftQuery

get_tutor

Returns the tutor with the given id

get_tutors

Returns an array of all the tutors in the databse

insert_shift_merging_if_needed 🔒

Inserts a new shift into the database, merging with any shifts that overlap

min_tutors_on_shift_in_interval_ignoring_a_specific_tutor 🔒

new_shift

Inserts a new shift into the database

new_tutor

Adds a tutor to the databse

patch_cleaning

Updates the cleaning for the shift with the given id

patch_tutor

Updates and returns the tutor with the given id

request_cover

Requests a cover for the shift with the given id

Type Aliases

SharedConnection 🔒

We share a single connection between threads. All the database manipulations happen through the DbConnection struct. In order to call those functions, we have to aquire the lock with .lock().await.