Struct csil_schedule_api::db::DbConnection

pub struct DbConnection(PgConnection);

This struct represents a connection to the database

A single DbConnection is created at startup and gets shared behind a tokio::sync::Mutex. Right now, the only supported operations are gets, inserts and updates. Because we don’t support more complicated queries, we sometimes make more round trips to the database than we need. Maybe consider adding more functions.

Using an async connection makes no difference because we share the connection between threads anyway.

Tuple Fields
0: PgConnection
Implementations

impl DbConnection

pub fn new() -> Self

Creates a new DbConnection from the DATABASE_URL env variable

This function also runs any pending migrations.

pub fn get_tutors(&mut self) -> Result<Vec<Tutor>, DbError>

Retrieves all the tutors in the database

pub fn get_tutor(&mut self, id: Uuid) -> Result<Option<Tutor>, DbError>

Retrieves the tutor with the given id

Returns None if there is no tutor in the databse with the given id.

pub fn insert_tutor(&mut self, tutor: &Tutor) -> Result<Tutor, DbError>

Inserts a new tutor into the database

Returns the tutor that was created.

pub fn update_tutor(&mut self, tutor: &Tutor) -> Result<Option<Tutor>, DbError>

Updates an existing tutor in the database

Returns the updated tutor or None if no tutor matches the id.

pub fn get_shift(&mut self, id: Uuid) -> Result<Option<Shift>, DbError>

Retrieves the shift with the given id

Returns None if there is no shift in the databse with the given id.

pub fn get_shifts(&mut self, query: ShiftQuery) -> Result<Vec<Shift>, DbError>

Retrieves the shifts that match the filters

pub fn insert_shift(&mut self, shift: &Shift) -> Result<Shift, DbError>

Inserts a new shift into the database

Returns the shift that was created.

pub fn update_shift(&mut self, shift: &Shift) -> Result<Option<Shift>, DbError>

Updates an existing shift in the database

Returns the updated shift or None if no shift matches the id.

pub fn delete_shift(&mut self, shift_id: Uuid) -> Result<Option<Shift>, DbError>

Deletes an existing shift in the database

Returns the shift or None if no shift matches the id.