Struct csil_schedule_api::types::DbShift

pub struct DbShift {
    pub id: Uuid,
    pub date: NaiveDate,
    pub start_time: i32,
    pub end_time: i32,
    pub tutor_id: Uuid,
    pub cleaning: Vec<Option<String>>,
    pub requesting_cover: bool,
}

A CSIL shift in the database

Conversions Into this type check that the date is in the valid range, the start and end times are in the valid range, the start and end times multiples of MINUTE_INCREMENT and the start time is stricly before the end time. This ensure that every shift is valid before it is put into the database.

In principle, this type can diverge from the Shift representation that the frontend and API use. For now, we just convert the [ChicagoTime]s into i32s and the [CleaningLocation]s into machine-readable Strings.

Fields
id: Uuiddate: NaiveDatestart_time: i32end_time: i32tutor_id: Uuidcleaning: Vec<Option<String>>requesting_cover: bool